st379 IMDI Fix People included in resources are not being exported
This appears to be an issue of users not understanding how the program
should work. All people should be entered first in the people tab. Then
they are included in the resource tab for the session filling different
roles. If this is done they are exported as actors and their roles are
listed.

card: https://trello.com/c/Xqwpex1h