Installer warning cleanup, new generated GUIDs, etc.
Note: The changes to the .guids... files for the sample data were generated before I realized that the sample data is no longer included in the Installer. I'm going ahead and checking them in just in case we were ever to go back and fix up the sample to be re-included in the Installer.