SP-995, SP-2186/SP-2265, SP-2144: Fixed various scenarios where, despite validation errors, SayMore would allow the user to go ahead with changing to a different view, task, etc. that would lead to an exception: Operation did not succeed because the program cannot commit or quit a cell value change.
Note that I was not able to reproduce SP-2186/SP-2265 in the current version, but I did determine that the previous fix for SP-2186 was actually for a different scenario that caused the same exception. I suspect that some subsequent change actually fixed the problem, but even if not, it's possible that this fix will.